Fix race condition in LifoBlockingLimiter
If a token was acquired around the same time that an operation
timed out, then the token would get lost. That means that the
limiter's capacity would effectively be reduced by one. If
this happens enough times, or the limit is decreased enough
for other reasons, then this would cause to the limiter to
permanently not let any operations through anymore.